GG-36023 Minor test refactoring.
