+++ modules/clients/src/test/java/org/apache/ignite/jdbc/thin/JdbcThinQueryMemoryTrackerSelfTest.java
@@ -21,20 +21,11 @@ import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import org.apache.ignite.IgniteCheckedException;
 import org.apache.ignite.configuration.IgniteConfiguration;
-import org.apache.ignite.internal.IgniteInternalFuture;
-import org.apache.ignite.internal.processors.cache.query.IgniteQueryErrorCode;
-import org.apache.ignite.internal.processors.query.oom.QueryMemoryTrackerSelfTest;
-import org.apache.ignite.internal.util.IgniteUtils;
-import org.apache.ignite.internal.util.typedef.X;
-import org.apache.ignite.testframework.GridTestUtils;
+import org.apache.ignite.internal.jdbc2.JdbcQueryMemoryTrackerSelfTest;
+import org.apache.ignite.internal.processors.query.h2.QueryMemoryTracker;
 import org.apache.ignite.testframework.ListeningTestLogger;
 import org.apache.ignite.testframework.LogListener;
-import org.apache.ignite.testframework.junits.GridAbstractTest;
 import org.junit.Test;
 
 import static org.apache.ignite.internal.util.IgniteUtils.MB;
@@ -42,85 +33,18 @@ import static org.apache.ignite.internal.util.IgniteUtils.MB;
 /**
  * Query memory manager for local queries.
  */
-public class JdbcThinQueryMemoryTrackerSelfTest extends QueryMemoryTrackerSelfTest {
+public class JdbcThinQueryMemoryTrackerSelfTest extends JdbcQueryMemoryTrackerSelfTest {
     /** Log listener. */
-    private static final LogListener logLsnr = LogListener.matches("Memory tracker has been closed concurrently").build();
+    private static final LogListener errMsgLsnr =
+        LogListener.matches(QueryMemoryTracker.ERROR_TRACKER_ALREADY_CLOSED).build();
 
     /** {@inheritDoc} */
     @Override protected IgniteConfiguration getConfiguration(String igniteInstanceName) throws Exception {
         ListeningTestLogger lnsrLog = new ListeningTestLogger(log());
 
-        lnsrLog.registerListener(logLsnr);
+        lnsrLog.registerListener(errMsgLsnr);
 
         return super.getConfiguration(igniteInstanceName).setGridLogger(lnsrLog);
-    }
-
-    /** {@inheritDoc} */
-    @Override protected boolean startClient() {
-        return false;
-    }
-
-    /** {@inheritDoc} */
-    @Test
-    @Override public void testGlobalQuota() throws Exception {
-        maxMem = 8 * MB;
-        useJdbcV2GlobalQuotaCfg = true;
-
-        final List<ResultSet> results = Collections.synchronizedList(new ArrayList<>());
-        final List<Statement> statements = Collections.synchronizedList(new ArrayList<>());
-        final List<IgniteInternalFuture> futs = new ArrayList<>();
-
-        try {
-            for (int i = 0; i < 10; i++) {
-                futs.add(GridTestUtils.runAsync(() -> {
-                        Connection conn = createConnection(true);
-                        Statement stmt = conn.createStatement();
-
-                        statements.add(stmt);
-
-                        ResultSet rs = stmt.executeQuery("select * from T as T0, T as T1 where T0.id < 2 " +
-                            "UNION select * from T as T2, T as T3 where T2.id >= 1 AND T2.id < 2");
-
-                        results.add(rs);
-
-                        rs.next();
-
-                        return null;
-                    }
-                ));
-            }
-
-            SQLException ex = (SQLException)GridTestUtils.assertThrows(GridAbstractTest.log, () -> {
-                SQLException sqlEx = null;
-
-                for (IgniteInternalFuture f : futs) {
-                    try {
-                        f.get(5_000);
-                    }
-                    catch (IgniteCheckedException e) {
-                        if (e.hasCause(SQLException.class))
-                            sqlEx = e.getCause(SQLException.class);
-                    }
-                }
-
-                if (sqlEx != null)
-                    throw sqlEx;
-
-                return null;
-            }, SQLException.class, "SQL query ran out of memory: Global quota was exceeded.");
-
-            assertEquals(IgniteQueryErrorCode.QUERY_OUT_OF_MEMORY, ex.getErrorCode());
-            assertEquals(IgniteQueryErrorCode.codeToSqlState(IgniteQueryErrorCode.QUERY_OUT_OF_MEMORY), ex.getSQLState());
-        }
-        finally {
-            for (ResultSet rs : results)
-                IgniteUtils.closeQuiet(rs);
-
-            for (Statement stmt : statements) {
-                IgniteUtils.closeQuiet(stmt.getConnection());
-                IgniteUtils.closeQuiet(stmt);
-            }
-        }
     }
 
     /**
@@ -129,8 +53,10 @@ public class JdbcThinQueryMemoryTrackerSelfTest extends QueryMemoryTrackerSelfTe
      * @throws Exception If failed.
      */
     @Test
-    public void testPartialRead() throws Exception {
+    public void testPartialResultRead() throws Exception {
         maxMem = 8 * MB;
+
+        System.out.println(">xxx> start test");
 
         try (Connection conn = createConnection(false)) {
             try (Statement stmt = conn.createStatement()) {
@@ -143,47 +69,11 @@ public class JdbcThinQueryMemoryTrackerSelfTest extends QueryMemoryTrackerSelfTe
             }
         }
 
-        assertFalse(logLsnr.check(1_000));
+        assertFalse(errMsgLsnr.check(1_000));
     }
 
     /** {@inheritDoc} */
-    @Override protected List<List<?>> execQuery(String sql, boolean lazy) throws Exception {
-        try (Connection conn = createConnection(lazy)) {
-            try (Statement stmt = conn.createStatement()) {
-                assert stmt != null && !stmt.isClosed();
-
-                try (ResultSet rs = stmt.executeQuery(sql)) {
-                    while (rs.next())
-                        ;
-                }
-            }
-        }
-        return Collections.emptyList();
-    }
-
-    /** {@inheritDoc} */
-    @Override protected void checkQueryExpectOOM(String sql, boolean lazy) {
-        try {
-            execQuery(sql, lazy);
-
-            fail("Exception is not thrown.");
-        } catch (SQLException e) {
-            assertTrue(e.getMessage().contains("SQL query ran out of memory: Query quota was exceeded."));
-            assertEquals(IgniteQueryErrorCode.QUERY_OUT_OF_MEMORY, e.getErrorCode());
-            assertEquals(IgniteQueryErrorCode.codeToSqlState(IgniteQueryErrorCode.QUERY_OUT_OF_MEMORY), e.getSQLState());
-        }
-        catch (Exception e) {
-            fail("Wrong exception: " + X.getFullStackTrace(e));
-        }
-    }
-
-    /**
-     * Initialize SQL connection.
-     *
-     * @param lazy Lazy flag.
-     * @throws SQLException If failed.
-     */
-    protected Connection createConnection(boolean lazy) throws SQLException {
+    @Override protected Connection createConnection(boolean lazy) throws SQLException {
         Connection conn = DriverManager.getConnection("jdbc:ignite:thin://127.0.0.1:10800..10802?" +
             "queryMaxMemory=" + (maxMem) + "&lazy=" + lazy);
